Many Changes:  - Fixed seq slide2slide to properly return a randomly sampled sliding window - with variable number of tokens predicted - moved record stats to bottom of custom Dataset.py -  fixed up LRScheduler with casting to float 32, Got altered version of positional encoding from chat GPT - Removed callbacks.py - no need to implement manually - Altered the name of multihead attention layer, conflicted with inbuilt Keras module - Modified inference.py to try and work with different saved model format from Model Checkpoint (Still not working) - Altred sample storage directory structure - Set up logging for training (Still needs fixes) - For train.py manually implemented early stopping - Added data.repeat in train_v2 to make sure we don’t accidentally occur the end of the dataset during an epoch - Removed custom dataset V2 - no need - Modified gitignore - Other changes in neural_translation not impt, from tutorial

+++ PyModel/Transformer/LRSchedule.py
@@ -13,7 +13,7 @@ class LRScheduler(LearningRateSchedule):
     def __call__(self, step_num):
         # print(step_num)
         # Linearly increasing the learning rate for the first warmup_steps, and decreasing it thereafter
-        arg1 = step_num ** -0.5
-        arg2 = step_num * (self.warmup_steps ** -1.5)
+        arg1 = tf.cast(step_num,tf.float32) ** tf.cast(-0.5, tf.float32)
+        arg2 = tf.cast(step_num, tf.float32) * tf.cast((self.warmup_steps ** -1.5), tf.float32)
  
         return tf.cast((self.d_model ** -0.5) * math.minimum(arg1, arg2), tf.float32)
